Reallocate memory block using malloc and free

+++ 0x0C-more_malloc_free/100-realloc.c
@@ -23,7 +23,7 @@ void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 {
 	void *p1;
 	char *p_old_char, *p_new_char;
-	int Q;
+	size_t Q;
 
 	if (ptr == NULL)
 		return (malloc(new_size));
